Record and initialize route
Filled out the POST and GET routes for records.

POST Route now creates a file on the file system and adds the
appropriate record entry in the record table.

Get Route returns the record information for a given record id

Added an initialization route that creates the actual database.

Also added a restclient file to test the record routes.