The big reconfigure
The catalog subsystem was original a folder within the larger
archivist repository. Organizing things this way was a mistake so I've
extracted the branch that had the catalog development work and moved
it into a new repository. All commits prior to this one come from the
archivist git history and were pushed to the empty repository. This
commit moves all of the catalog level files into the root of the new
repository.